+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [2.2.0] - 2024-10-29
+
+### 2.2.0 Added
+
+- Added ProteinMPNN to RFDiffusion workflow
+- Added higher-level nanobody design recipe
+
+---
+
 ## [2.1.1] - 2024-10-15
 
 ### 2.1.1 Changed